requirements.txt
Django==3.0.2
mysqlclient==1.4.6
djangorestframework-jwt==1.11.0